sending ethers and variables updated, bash script completed
