+++ example_code/basic_examples/variables/Cargo.toml
@@ -1,15 +1,35 @@
 [package]
-name = "events"
-version = "0.1.0"
+name = "stylus_variable_example"
+version = "0.1.7"
 edition = "2021"
+license = "MIT OR Apache-2.0"
+homepage = "https://github.com/OffchainLabs/stylus-hello-world"
+repository = "https://github.com/OffchainLabs/stylus-hello-world"
+keywords = ["arbitrum", "ethereum", "stylus", "alloy"]
+description = "Stylus hello world example"
 
 [dependencies]
-stylus-sdk = { version = "0.4.2", features = ["debug"] }
-wee_alloc = "0.4.5"
-alloy-sol-types = "0.3.1"
+alloy-primitives = "0.7.6"
+alloy-sol-types = "0.7.6"
+mini-alloc = "0.4.2"
+stylus-sdk = "0.5.2"
+hex = "0.4.3"
+
+[dev-dependencies]
+tokio = { version = "1.12.0", features = ["full"] }
+ethers = "2.0"
+eyre = "0.6.8"
 
 [features]
 export-abi = ["stylus-sdk/export-abi"]
+debug = ["stylus-sdk/debug"]
+
+[[bin]]
+name = "stylus-hello-world"
+path = "src/main.rs"
+
+[lib]
+crate-type = ["lib", "cdylib"]
 
 [profile.release]
 codegen-units = 1
@@ -17,5 +37,3 @@ strip = true
 lto = true
 panic = "abort"
 opt-level = "s"
-
-[workspace]
