+++ example_code/basic_examples/variables/rust-toolchain.toml
@@ -1,0 +1,2 @@
+[toolchain]
+channel = "1.79.0"
