+++ example_code/basic_examples/variables/src/main.rs
@@ -1,59 +1,4 @@
-#![no_main]
-#![no_std]
-extern crate alloc;
-
-#[global_allocator]
-static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-use alloc::vec;
-use alloc::vec::Vec;
-
-use stylus_sdk::alloy_primitives::{U16, U256};
-use stylus_sdk::prelude::*;
-use stylus_sdk::storage::{StorageAddress, StorageBool, StorageU256};
-use stylus_sdk::{block, console, msg};
-
-#[solidity_storage]
-#[entrypoint]
-pub struct Contract {
-    initialized: StorageBool,
-    owner: StorageAddress,
-    max_supply: StorageU256,
-}
-
-#[external]
-impl Contract {
-    // State variables are initialized in an `init` function.
-    pub fn init(&mut self) -> Result<(), Vec<u8>> {
-        // We check if contract has been initialized before.
-        // We return if so, we initialize if not.
-        let initialized = self.initialized.get();
-        if initialized {
-            return Ok(());
-        }
-        self.initialized.set(true);
-
-        // We set the contract owner to the caller,
-        // which we get from the global msg module
-        self.owner.set(msg::sender());
-        self.max_supply.set(U256::from(10_000));
-
-        Ok(())
-    }
-
-    pub fn do_something() -> Result<(), Vec<u8>> {
-        // Local variables are not saved to the blockchain
-        // 16-bit Rust integer
-        let i = 456_u16;
-        // 16-bit int inferred from U16 Alloy primitive
-        let j = U16::from(123);
-
-        // Here are some global variables
-        let timestamp = block::timestamp();
-        let amount = msg::value();
-
-        console!("Local variables: {i}, {j}");
-        console!("Global variables: {timestamp}, {amount}");
-
-        Ok(())
-    }
+#[cfg(feature = "export-abi")]
+fn main() {
+    stylus_variable_example::print_abi("MIT-OR-APACHE-2.0", "pragma solidity ^0.8.23;");
 }
